expansion to ai category
