+++ extern/util/static.js
@@ -372,6 +372,10 @@ EntryStatic.getAllBlocks = function() {
             blocks: ['functionAddButton'],
         },
         {
+            category: 'artificial_intelligence',
+            blocks: ['audio_title', 'check_microphone', 'speech_to_text', 'get_microphone_volume'],
+        },
+        {
             category: 'expansion',
             blocks: [
                 'expansionBlockAddButton',
@@ -402,10 +406,6 @@ EntryStatic.getAllBlocks = function() {
                 'read_text',
                 'read_text_wait_with_block',
                 'set_tts_property',
-                'audio_title',
-                'check_microphone',
-                'speech_to_text',
-                'get_microphone_volume',
             ],
         },
         {
@@ -785,6 +785,7 @@ EntryStatic.colorSet = {
             VARIABLE: '#CE38CE',
             HARDWARE: '#FFFFFF',
             EXPANSION: '#FF8888',
+            ARTIFICIAL_INTELLIGENCE: '#FF8888',
             HIDDEN: '#FFFFFF',
         },
     },
@@ -803,6 +804,7 @@ EntryStatic.colorSet = {
             FUNC: '#de5c04',
             HARDWARE: '#00b6b1',
             EXPANSION: '#ef6d6d',
+            ARTIFICIAL_INTELLIGENCE: '#8222ff',
             HIDDEN: '#8aa3b2',
         },
         lighten: {
@@ -819,6 +821,7 @@ EntryStatic.colorSet = {
             FUNC: '#ff7b22',
             HARDWARE: '#78d5d3',
             EXPANSION: '#ffaeae',
+            ARTIFICIAL_INTELLIGENCE: '#ffaeae',
             HIDDEN: '#ffaeae',
         },
         darken: {
@@ -835,6 +838,7 @@ EntryStatic.colorSet = {
             FUNC: '#a14100',
             HARDWARE: '#008380',
             EXPANSION: '#c63f3f',
+            ARTIFICIAL_INTELLIGENCE: '#670bdd',
             HIDDEN: '#728997',
         },
         emphasize: {
