+++ src/playground/blocks/index.js
@@ -17,6 +17,7 @@ const basicBlockList = [
     require('./block_calc'),
     require('./block_variable'),
     require('./block_func'),
+    require('./block_artificial_intelligence'),
 ];
 
 Entry.EXPANSION_BLOCK = {};
@@ -26,7 +27,6 @@ require('./block_expansion_translate');
 require('./block_expansion_behaviorconduct_disaster');
 require('./block_expansion_behaviorconduct_lifesafety');
 require('./block_expansion_tts');
-require('./block_expansion_audio');
 Entry.EXPANSION_BLOCK_LIST = {
     weather: Entry.Expansion_Weather,
     festival: Entry.EXPANSION_BLOCK.festival,
@@ -34,7 +34,6 @@ Entry.EXPANSION_BLOCK_LIST = {
     behaviorConductDisaster: Entry.EXPANSION_BLOCK.behaviorConductDisaster,
     behaviorConductLifeSafety: Entry.EXPANSION_BLOCK.behaviorConductLifeSafety,
     tts: Entry.EXPANSION_BLOCK.tts,
-    audio: Entry.EXPANSION_BLOCK.audio,
 };
 
 function getBlockObject(items) {
